+++ Clients/src/presentation/components/AddNewRiskForm/index.tsx
@@ -7,6 +7,9 @@ import "./styles.module.css";
 import { Likelihood, Severity } from "../RiskLevel/constants";
 import { RiskFormValues, RiskFormErrors, MitigationFormValues, MitigationFormErrors } from "./interface";
 
+import { checkStringValidation } from "../../../application/validations/stringValidation";
+import selectValidation from "../../../application/validations/selectValidation";
+
 const RiskSection = lazy(() => import("./RisksSection"));
 const MitigationSection = lazy(() => import("./MitigationSection"));
 
@@ -69,6 +72,8 @@ const AddNewRiskForm: FC<AddNewRiskFormProps> = ({
   const disableRipple =
     theme.components?.MuiButton?.defaultProps?.disableRipple;
   
+  const [riskErrors, setRiskErrors] = useState<RiskFormErrors>({});
+  const [migitateErrors, setMigitateErrors] = useState<MitigationFormErrors>({});
   const [riskValues, setRiskValues] = useState<RiskFormValues>(riskInitialState);
   const [mitigationValues, setMitigationValues] = useState<MitigationFormValues>(mitigationInitialState);
   const [value, setValue] = useState("risks");
@@ -90,6 +95,142 @@ const AddNewRiskForm: FC<AddNewRiskFormProps> = ({
     }),
     []
   );
+
+  const validateForm = useCallback((): boolean => {
+    const newErrors: RiskFormErrors = {};
+    const newMitigationErrors: MitigationFormErrors = {};
+
+    const riskName = checkStringValidation("Risk name", riskValues.riskName, 3, 50);
+    if (!riskName.accepted) {
+      newErrors.riskName = riskName.message;
+    }
+    const riskDescription = checkStringValidation(
+      "Risk description",
+      riskValues.riskDescription,
+      1,
+      256
+    );
+    if (!riskDescription.accepted) {
+      newErrors.riskDescription = riskDescription.message;
+    }
+    const potentialImpact = checkStringValidation(
+      "Potential impact",
+      riskValues.potentialImpact,
+      1,
+      256
+    );
+    if (!potentialImpact.accepted) {
+      newErrors.potentialImpact = potentialImpact.message;
+    }
+    const reviewNotes = checkStringValidation(
+      "Review notes",
+      riskValues.reviewNotes,
+      0,
+      1024
+    );
+    if (!reviewNotes.accepted) {
+      newErrors.reviewNotes = reviewNotes.message;
+    }
+    const actionOwner = selectValidation("Action owner", riskValues.actionOwner);
+    if (!actionOwner.accepted) {
+      newErrors.actionOwner = actionOwner.message;
+    }
+    const aiLifecyclePhase = selectValidation(
+      "AI lifecycle phase",
+      riskValues.aiLifecyclePhase
+    );
+    if (!aiLifecyclePhase.accepted) {
+      newErrors.aiLifecyclePhase = aiLifecyclePhase.message;
+    }
+    const riskCategory = selectValidation("Risk category", riskValues.riskCategory);
+    if (!riskCategory.accepted) {
+      newErrors.riskCategory = riskCategory.message;
+    }
+
+    const mitigationPlan = checkStringValidation(
+      "Mitigation plan",
+      mitigationValues.mitigationPlan,
+      1,
+      1024
+    );
+    if (!mitigationPlan.accepted) {
+      newMitigationErrors.mitigationPlan = mitigationPlan.message;
+    }
+    const implementationStrategy = checkStringValidation(
+      "Implementation strategy",
+      mitigationValues.implementationStrategy,
+      1,
+      1024
+    );
+    if (!implementationStrategy.accepted) {
+      newMitigationErrors.implementationStrategy = implementationStrategy.message;
+    }
+    // const recommendations = checkStringValidation(
+    //   "Recommendations",
+    //   mitigationValues.recommendations,
+    //   1,
+    //   1024
+    // );
+    // if (!recommendations.accepted) {
+    //   newMitigationErrors.recommendations = recommendations.message;
+    // }
+    const deadline = checkStringValidation(
+      "Deadline",
+      mitigationValues.deadline,
+      1
+    );
+    if (!deadline.accepted) {
+      newMitigationErrors.deadline = deadline.message;
+    }
+    const dateOfAssessment = checkStringValidation(
+      "Date Of Assessment",
+      mitigationValues.dateOfAssessment,
+      1
+    );
+    if (!dateOfAssessment.accepted) {
+      newMitigationErrors.dateOfAssessment = dateOfAssessment.message;
+    }
+    const mitigationStatus = selectValidation(
+      "Mitigation status",
+      mitigationValues.mitigationStatus
+    );
+    if (!mitigationStatus.accepted) {
+      newMitigationErrors.mitigationStatus = mitigationStatus.message;
+    }
+    const currentRiskLevel = selectValidation(
+      "Current risk level",
+      mitigationValues.currentRiskLevel
+    );
+    if (!currentRiskLevel.accepted) {
+      newMitigationErrors.currentRiskLevel = currentRiskLevel.message;
+    }
+    const approver = selectValidation("Approver", mitigationValues.approver);
+    if (!approver.accepted) {
+      newMitigationErrors.approver = approver.message;
+    }
+    const approvalStatus = selectValidation(
+      "Approval status",
+      mitigationValues.approvalStatus
+    );
+    if (!approvalStatus.accepted) {
+      newMitigationErrors.approvalStatus = approvalStatus.message;
+    }
+
+    setMigitateErrors(newMitigationErrors);
+    setRiskErrors(newErrors);
+
+    return Object.keys(newErrors).length === 0 && Object.keys(newMitigationErrors).length === 0;  // Return true if no errors exist    
+  }, [riskValues, mitigationValues]);
+
+  const riskFormSubmitHandler = () => {
+    // check forms validate    
+    if(validateForm()){
+      // call backend 
+      closePopup();
+    }else{
+      console.log('validation fails')
+    }
+  }
 
   return (
     <Stack>
@@ -119,15 +260,16 @@ const AddNewRiskForm: FC<AddNewRiskFormProps> = ({
         </Box>
         <Suspense fallback={<div>Loading...</div>}>
           <TabPanel value="risks" sx={{ p: "24px 0 0" }}>
-            <RiskSection closePopup={closePopup} riskValues={riskValues} setRiskValues={setRiskValues}/>
+            <RiskSection riskValues={riskValues} setRiskValues={setRiskValues} riskErrors={riskErrors}/>
           </TabPanel>
           <TabPanel value="mitigation" sx={{ p: "24px 0 0" }}>
-            <MitigationSection closePopup={closePopup} mitigationValues={mitigationValues} setMitigationValues={setMitigationValues} />
+            <MitigationSection mitigationValues={mitigationValues} setMitigationValues={setMitigationValues} migitateErrors={migitateErrors} />
           </TabPanel>
         </Suspense>
         <Box sx={{ display: 'flex'}}>
           <Button
-            type="button"            
+            type="submit" 
+            onClick={riskFormSubmitHandler}           
             variant="contained"
             disableRipple={
               theme.components?.MuiButton?.defaultProps?.disableRipple
